# applications/oxybank/web/openapi/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/kb_base": {
      "get": {
        "tags": [
          "Knowledge Base Management"
        ],
        "summary": "Get all knowledge bases",
        "description": "Returns a list of all knowledge bases in the system",
        "operationId": "get_all_knowledge_base_api_v1_kb_base_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number, starting from 1",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number, starting from 1"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Page size, options: 10/20/50/100/200",
              "default": 10,
              "title": "Size"
            },
            "description": "Page size, options: 10/20/50/100/200"
          }
        ],
        "responses": {
          "200": {
            "description": "Knowledge base list containing ID, name, description, type, path, and other information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_PaginatedResponse_KnowledgeBaseItem__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Knowledge Base Management"
        ],
        "summary": "Create an empty knowledge base",
        "description": "Create an empty knowledge base",
        "operationId": "create_knowledge_base_api_v1_kb_base_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_str_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kb_base/{kb_name}": {
      "delete": {
        "tags": [
          "Knowledge Base Management"
        ],
        "summary": "Delete knowledge base",
        "description": "Delete the specified knowledge base and its indexes and vector space",
        "operationId": "delete_knowledge_base_api_v1_kb_base__kb_name__delete",
        "parameters": [
          {
            "name": "kb_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Knowledge base name",
              "title": "Kb Name"
            },
            "description": "Knowledge base name"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_str_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kb_base/{kb_name}/schema": {
      "post": {
        "tags": [
          "Knowledge Base Management"
        ],
        "summary": "Update knowledge base Schema",
        "description": "Update the Schema configuration of the specified knowledge base",
        "operationId": "update_kb_schema_api_v1_kb_base__kb_name__schema_post",
        "parameters": [
          {
            "name": "kb_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Knowledge base name",
              "title": "Kb Name"
            },
            "description": "Knowledge base name"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KBSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_str_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kb_base/{kb_name}/schema/exists": {
      "get": {
        "tags": [
          "Knowledge Base Management"
        ],
        "summary": "Check if knowledge base Schema exists",
        "description": "Check if the specified knowledge base has Schema configured",
        "operationId": "get_kb_schema_api_v1_kb_base__kb_name__schema_exists_get",
        "parameters": [
          {
            "name": "kb_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Knowledge base name",
              "title": "Kb Name"
            },
            "description": "Knowledge base name"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_bool_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_PaginatedResponse_KnowledgeBaseItem__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Response status code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "Response message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginatedResponse_KnowledgeBaseItem_"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response data"
          }
        },
        "type": "object",
        "title": "APIResponse[PaginatedResponse[KnowledgeBaseItem]]"
      },
...
      "APIResponse_bool_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Response status code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "Response message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response data"
          }
        },
        "type": "object",
        "title": "APIResponse[bool]"
      },
...
      "APIResponse_str_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Response status code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "Response message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response data"
          }
        },
        "type": "object",
        "title": "APIResponse[str]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KBSchema": {
        "properties": {
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldInfo"
            },
            "type": "array",
            "title": "Fields"
          },
          "match_rules": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MatchRule"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Rules"
          },
          "parser_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParserConfig"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "fields"
        ],
        "title": "KBSchema"
      },
      "KnowledgeBaseItem": {
        "properties": {
          "kb_id": {
            "type": "string",
            "title": "Kb Id",
            "description": "Knowledge base ID",
            "default": "",
            "example": "d1686c75272e7ab78643367eb438751c"
          },
          "kb_name": {
            "type": "string",
            "title": "Kb Name",
            "description": "Knowledge base name",
            "example": "kb_test1"
          },
          "kb_type": {
            "type": "string",
            "title": "Kb Type",
            "description": "Knowledge base type",
            "example": "unstructured"
          },
          "kb_description": {
            "type": "string",
            "title": "Kb Description",
            "description": "Knowledge base description",
            "default": "",
            "example": "this is kb kb_test1, file_path is /Users/yulei116/workspace/OxyGentKnowledgeBank-1/kb_data/kb_test1, kb_id is c1caa264-ce70-4544-8ffa-108dac1d6f64"
          },
          "kb_status": {
            "type": "string",
            "title": "Kb Status",
            "description": "Knowledge base status",
            "default": "active",
            "example": "active"
          },
          "create_time": {
            "type": "string",
            "title": "Create Time",
            "description": "Knowledge base creation time",
            "default": "",
            "example": "2025-01-01 10:00:00"
          },
          "update_time": {
            "type": "string",
            "title": "Update Time",
            "description": "Knowledge base update time",
            "default": "",
            "example": "2025-01-01 10:00:00"
          },
          "kb_create_user": {
            "type": "string",
            "title": "Kb Create User",
            "description": "Knowledge base creator",
            "default": "",
            "example": "bob"
          },
          "kb_update_user": {
            "type": "string",
            "title": "Kb Update User",
            "description": "Knowledge base updater",
            "default": "",
            "example": "bob"
          },
          "kb_store_type": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Kb Store Type",
            "description": "Knowledge base storage type",
            "example": "[elasticsearch]"
          },
          "kb_extra_info": {
            "additionalProperties": true,
            "type": "object",
            "title": "Kb Extra Info",
            "description": "Knowledge base extra information"
          },
          "kb_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Kb Schema",
            "description": "Knowledge base schema information"
          },
          "auto_bind_query": {
            "type": "boolean",
            "title": "Auto Bind Query",
            "description": "Whether to automatically bind query interfaces for retrieval strategies on restart",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "kb_name",
          "kb_type"
        ],
        "title": "KnowledgeBaseItem",
        "description": "Knowledge base item model"
      },
...
    }
  }
}